영화 리스트 페이지 => 상세 페이지 => 수정 페이지
