+++ frontend/src/Pages/Admin/index.tsx
@@ -1,8 +1,9 @@
 export { default as AdminMovieExact } from './AdminMovieExact';
+export { default as AdminModifyMovie } from './AdminModifyMovie';
 export { default as AdminHome } from './AdminHome';
 export { default as AdminTheater } from './AdminTheater';
 export { default as AdminHall } from './AdminHall';
-export { default as AdminMovie } from './AdminMovie';
+export { default as AdminMovieList } from './AdminMovieList';
 export { default as AdminPeople } from './AdminPeople';
 export { default as AdminShowSCH } from './AdminShowSCH';
 export { default as AdminSales } from './AdminSales';
